+++ scripts/api_test.py
@@ -6,27 +6,31 @@ curl http://localhost:8080/ --include --header "Content-Type: application/json" 
 """
 
 
+# from util_req import
+
+# from ref_types.transactions import query_tx
 from util_base import API_URL
-from util_req import RequestBase, RequestType, send_request
-from util_txs import query_tx
+
+# from util_txs import query_tx
+
 
 CHAIN_ID = "localjuno-1"
-bin_base = RequestBase(API_URL, CHAIN_ID, RequestType.BIN)
-query_base = RequestBase(API_URL, CHAIN_ID, RequestType.QUERY)
+# bin_base = RequestBase(API_URL, CHAIN_ID, RequestType.BIN)
+# query_base = RequestBase(API_URL, CHAIN_ID, RequestType.QUERY)
 
 from ref_types.transactions import RequestBuilder
 
+# TODO: option to toggle printing in send_request for all? (off by default)
+rb = RequestBuilder(API_URL, CHAIN_ID)
+
 
 def main():
-    bin_test()
-    # tx_test()
+    # bin_test()
+    tx_test()
 
 
 # Test to ensure the base layer works and returns data properly
 def bin_test():
-    # TODO: option to toggle printing in send_request for all? (off by default)
-    rb = RequestBuilder(API_URL, CHAIN_ID)
-
     rb.bin("keys list --keyring-backend=test --output=json")
     rb.bin(
         "tx decode ClMKUQobL2Nvc21vcy5nb3YudjFiZXRhMS5Nc2dWb3RlEjIIpwISK2p1bm8xZGM3a2MyZzVrZ2wycmdmZHllZGZ6MDl1YTlwZWo1eDNsODc3ZzcYARJmClAKRgofL2Nvc21vcy5jcnlwdG8uc2VjcDI1NmsxLlB1YktleRIjCiECxjGMmYp4MlxxfFWi9x4u+jOleJVde3Cru+HnxAVUJmgSBAoCCH8YNBISCgwKBXVqdW5vEgMyMDQQofwEGkDPE4dCQ4zUh6LIB9wqNXDBx+nMKtg0tEGiIYEH8xlw4H8dDQQStgAe6xFO7I/oYVSWwa2d9qUjs9qyB8r+V0Gy"
@@ -44,20 +48,17 @@ def bin_test():
 
 # Test to ensure Transactions and getting that data returns properly
 def tx_test():
-    res = send_request(
-        bin_base,
-        "tx bank send acc0 juno10r39fueph9fq7a6lgswu4zdsg8t3gxlq670lt0 500ujuno --fees 5000ujuno --node %RPC% --chain_id %CHAIN_ID% --yes --output json",
+    res = rb.bin(
+        "tx bank send acc0 juno10r39fueph9fq7a6lgswu4zdsg8t3gxlq670lt0 500ujuno --fees 5000ujuno --node %RPC% --chain-id %CHAIN_ID% --yes --output json --keyring-backend=test"
     )
-    print(res)
-
-    tx_data = query_tx(query_base, res)
+    tx_data = rb.query_tx(res)
     print(tx_data)
 
-    res = send_request(
-        query_base,
-        "bank balances juno10r39fueph9fq7a6lgswu4zdsg8t3gxlq670lt0 --output=json",
+    print(
+        rb.query(
+            "bank balances juno10r39fueph9fq7a6lgswu4zdsg8t3gxlq670lt0 --output=json"
+        )
     )
-    print(res)
 
 
 if __name__ == "__main__":
