+++ scripts/util_contracts.py
@@ -1,12 +1,12 @@
 import base64
 import json
 
+from ref_types.transactions import RequestBuilder, send_request
 from util_base import contracts_path, current_dir, parent_dir
-from util_req import RequestBase, send_request
 
 
-def get_contract_address(query_base: RequestBase, tx_hash: str) -> str:
-    res_json = send_request(query_base, f"tx {tx_hash} --output=json")
+def get_contract_address(rb: RequestBuilder, tx_hash: str) -> str:
+    res_json = rb.query(f"tx {tx_hash} --output=json")
 
     code = int(res_json["code"])
     if code != 0:
